[#1174] SbS: Move model classes out of the CompactBlockProcessor
